数据库类型和c#类型转换的时候, decimal不能当成double. 否则转换失败.

+++ JuBuilder/CsType.cs
@@ -17,7 +17,7 @@ namespace JuBuilder
             alias["nvarchar"] = "string";
             alias["datetime"] = "DateTime";
             alias["float"] = "double";
-            alias["decimal"] = "double";
+            alias["decimal"] = "decimal";
             alias["bit"] = "bool";
         }
         public string GetLower(string type)
